requirements.txt
faiss-cpu>=1.7.0
numpy>=1.23.0
pandas>=1.5.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pycryptodome>=3.15.0
opencv-python>=4.7.0
easyocr>=1.6.2
streamlit>=1.23.0
python-dotenv>=0.21.0
sentence-transformers>=2.2.2
transformers 
torch
pillow>=9.0.0
pytesseract>=0.3.9
librosa
flask>=2.2.0
flask-cors>=3.0.10
requests>=2.28.0
pymongo 
bcrypt
itsdangerous
youtube_transcript_api
pytube
yt_dlp
openai-whisper
ffmpeg-python
pyspellchecker